fix default contract and loop

+++ src/fractal-contract.cpp
@@ -284,7 +284,7 @@ void fractal_contract::submitcons(const uint64_t& groupnr, const std::vector<nam
 
     check(is_account(submitter), "Submitter's account does not exist.");
 
-     for (int i = 0; i < rankings.size(); i++)
+     for (size_t i = 0; i < rankings.size(); i++)
     {
         std::string rankname = rankings[i].to_string();
